Django/mi_proyecto/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
